`Tests`: Split PurchasesOrchestratorTests.swift (#3695)
`PurchasesOrchestratorTests.swift` was getting too big and hard to
understand.

I've made a first pass at splitting the tests in three files:

`PurchasesOrchestratorSK1Tests.swift`: Tests related to StoreKit 1
`PurchasesOrchestratorSK2Tests.swift`: Tests related to StoreKit 2
`PurchasesOrchestratorCommonTests.swift`: Tests related to both versions
of StoreKit

My plan is to further clean and organize the tests but I wanted to first
split them as they are, without changing any test.
